clean(category): remove categorizer system
This system helps to make the update process easier by automatically performing basic categories when Minecraft was updated and items were added. In fact, it helped a lot with the 1.18 and 1.19 updates.

It was also intended to reduce the workload of users by providing basic categories using the 'tag' system, even if the Box is no longer maintained.

However, considering future changes to the Spigot/Paper API such that Material is no longer an Enum, this system will either be very difficult to maintain or will require a major re-implementation. This was decided to remove as I felt that it was not really worth the work.