fix(core,storage): first aid for updating items that have no effects
In Minecraft 1.20.2, effects are no longer saved as integer. The paths for saving effects have been changed accordingly. If an ItemStack with no effects is loaded in 1.20.2, an NPE will occur, which itself cannot be fixed by the Box. This commit avoids the NPE by creating a new ItemStack for the causative item.

In the future, we will need to rethink the management of default items. This is because items are being created that should not exist, such as this one.