fix(stick): consuming arrows from Box
In 1.20.5, `EntityShootBowEvent#setConsumeItem` is no longer working. This is because Mojang changed the code to consuming arrows before the event fires.
Instead, I re-implemented it by re-adding arrows to the player.
Due to this change, the behavior when the arrow is only one in the inventory will change. The slot in which the arrow was located will no longer be kept as before. The arrow will be added to the first available slot.