change(stick): no longer calling Inventory open/click events
Spigot has converted InventoryView from an abstract class to an interface, so we need to implement all methods of that if we use custom InventoryView.
Since it is hard to maintain InventoryView any longer, we decided to give up firing events.
The integrations for WorldGuard and CoreProtect are directly implemented in the previous commits.